refactor: Users 일급 객체 만든 뒤, 수강인원 제한 기능 책임 부여 및 테스트 케이스 수정
