+++ src/main/java/nextstep/session/domain/Session.java
@@ -4,16 +4,12 @@ import nextstep.image.domain.Image;
 import nextstep.users.domain.NsUser;
 
 import java.time.LocalDateTime;
-import java.util.HashSet;
-import java.util.Set;
 
 public class Session {
 
     private Long id;
 
-    private Set<NsUser> members = new HashSet<>();
-
-    private int numberOfMaximumMembers;
+    private Users members;
 
     private SessionType sessionType;
 
@@ -25,10 +21,9 @@ public class Session {
 
     private EndAt endAt;
 
-    public Session(Long id, Set<NsUser> members, int numberOfMaximumMembers, SessionType sessionType, SessionStatus status, Image coverImage, StartAt startAt, EndAt endAt) {
+    public Session(Long id, Users members, SessionType sessionType, SessionStatus status, Image coverImage, StartAt startAt, EndAt endAt) {
         this.id = id;
         this.members = members;
-        this.numberOfMaximumMembers = numberOfMaximumMembers;
         this.sessionType = sessionType;
         this.status = status;
         this.coverImage = coverImage;
@@ -37,7 +32,7 @@ public class Session {
     }
 
     public static Session create(
-            int numberOfMaximumMembers,
+            Users members,
             SessionType sessionType,
             Image coverImage,
             LocalDateTime startAt,
@@ -45,8 +40,7 @@ public class Session {
     ) {
         return new Session(
                 null,
-                new HashSet<>(),
-                numberOfMaximumMembers,
+                members,
                 sessionType,
                 SessionStatus.PREPARING,
                 coverImage,
@@ -59,19 +53,10 @@ public class Session {
         if (status != SessionStatus.RECRUITING) {
             throw new IllegalStateException("수강신청은 모집중인 상태일 때만 가능합니다.");
         }
-
-        if (sessionType == SessionType.FREE) {
-            return;
-        }
-
-        if (numberOfMaximumMembers >= members.size()) {
-            throw new IllegalStateException("최대 수강 인원을 초과하였습니다.");
-        }
     }
 
     public void register(NsUser user) {
         isRegistrable();
-
-        members.add(user);
+        members.register(user, sessionType);
     }
 }
